simple_jarvis/requirements.txt
SpeechRecognition
pyttsx3
deep-translator
langdetect
huggingface_hub
python-dotenv
Pillow
pyaudio
